requirements.txt
Django==2.2.1
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==0.0.8
django-extensions==2.1.6
psycopg2==2.8.2
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
